attach items on services

+++ src/components/admin/Services.vue
@@ -1,11 +1,19 @@
 <script setup>
-import { reactive, ref, onMounted } from "vue";
+import { reactive, ref, onMounted, onBeforeMount  } from "vue";
 import { store } from "../../store/service";
+import Swal from 'sweetalert2';
 const service = store();
 
 
 const formDialog = ref(false)
 const Update = ref(false);
+const addServicesProductsDialog = ref(false);
+const selectedService = ref(null);
+
+const ServicesItems = reactive([{
+    product_id: null,
+    quantity: 0,
+}]);
 
 const addServices = () => {
     service.createServices().then(response => {
@@ -19,11 +27,13 @@ const addServices = () => {
 const closeDialog = () => {
     formDialog.value = false;
     Update.value = false;
+    addServicesProductsDialog.value = false;
     resetFields();
 }
 
 const resetFields = () => {
-    service.services = {}
+    service.services = {};
+    selectedService.value = null;
 }
 
 const paginate = (page) => {
@@ -31,7 +41,25 @@ const paginate = (page) => {
 }
 
 const deleteServices = (id) => {
-    service.deleteService(id)
+    Swal.fire({
+    title: 'Delete Service?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        'Deleted!',
+        'Services Successfully Deleted',
+        'success'
+      )
+      service.deleteService(id)
+    }
+  })
+   
 }
 
 const updateServices = (data) => {
@@ -52,8 +80,14 @@ const submitUpdate = () => {
 
 onMounted(() => {
     service.getServices();
+    service.getServicesProductDropdown();
 });
 
+onBeforeMount(() => {
+ service.getServices();
+ service.getServicesProductDropdown();
+});
+ 
 const formatPrice =(price) =>{
     return `₱${parseFloat(price).toLocaleString('en-US', {minimumFractionDigits: 2})}`;
 }
@@ -67,6 +101,48 @@ const isNumber = function(evt) {
   } else {
     return true;
   }
+}
+
+const addMoreItems = () => {
+    ServicesItems.push({
+        product_id: null,
+        quantity: 0,
+    });
+}
+
+const addServicesProduct = (service_id) => { 
+   selectedService.value = service_id 
+   addServicesProductsDialog.value = true;
+}
+
+const removeItem = (index) => {
+    ServicesItems.splice(index, 1);
+}
+
+const addServicesProducts = () =>{
+    service.attachProductsOnService(selectedService.value, ServicesItems).then(response => {
+        if(response.data.status == 'success'){
+            closeDialog();
+        }else{
+            closeDialog();
+        }  
+    });
+}
+
+const removeServicesProduct = (services_id) => {    
+    Swal.fire({
+        title: 'Remove Services Items?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+        if (result.isConfirmed) {
+        service.removeAttachedProducts(services_id);
+        }
+    })
 }
 
 </script>
@@ -84,21 +160,34 @@ const isNumber = function(evt) {
                   <tr>
                     <th scope="col">Service Name</th>
                     <th scope="col">Service Type</th>
-                    <th scope="col">Price</th>
+                    <th scope="col">Service Price</th>
                     <th scope="col">Details</th>
-                    <th scope="col">Action</th>
+                    <th scope="col">Package Included</th>
+                    <th scope="col">Total Price</th>
+                    <th scope="col" class="actions">Action</th>
                   </tr>
                 </thead>
-                <tbody v-for="(data,index) in service.services_details.data" :key="index">
+                <tbody v-for="(data,index) in service.services_details.data" :key="index" class="table-actions">
                   <tr>
                  
                     <td>{{ data.name }}</td>
                     <td>{{ data.type}}</td>
                     <td>{{ formatPrice(data.price) }}</td>
-                    <td>{{ data.details }}</td>
+                    <td >{{ data.details }} 
+                    </td>
+                    <td>
+                        <div v-for="(product,i) in data.products">
+                            {{ product.name }} {{ formatPrice(product.price) }} x {{ product.quantity }} pcs. = {{ formatPrice(product.price * product.quantity ) }}
+                        </div>
+                    </td>
                     <td> 
-                        <span @click="updateServices(data)"><i class="fa-solid fa-edit"></i></span>
-                        <span @click="deleteServices(data.id)"><i class="fa-solid fa-trash"></i></span>
+                        {{ formatPrice(data.price + data.total_product_price ) }}
+                    </td>
+                    <td class="actions d-flex flex-column" style="border: none;"> 
+                        <span @click="updateServices(data)">Edit Services</span>
+                        <span @click="deleteServices(data.id)">Delete Services</span>
+                        <span @click="addServicesProduct(data.id)">Add Items</span>
+                        <span @click="removeServicesProduct(data.id)">Remove Items</span>
                     </td>
 
                   </tr>
@@ -106,7 +195,7 @@ const isNumber = function(evt) {
 
               </table>
         
-              <div v-if="service.services_details.total > 10" class="table-pagination">  
+              <div v-if="service.services_details.total > 3" class="table-pagination">  
                 <div>
                     <button @click="paginate(service.services_details.current_page - 1)" v-if="service.services_details.prev_page_url"><i class="fa fa-angle-left"></i></button>
                     <button @click="paginate(service.services_details.current_page + 1)" v-if="service.services_details.next_page_url" ><i class="fa fa-angle-right"></i></button>
@@ -165,6 +254,63 @@ const isNumber = function(evt) {
               </form>
 
               <div v-if="Update">
+                <button class="btn btn-primary" @click="submitUpdate()">Update</button>
+                <button type="button" class="btn btn-danger" @click="closeDialog">Cancel</button>
+              </div>
+           
+            </div>
+            </div>
+        </div>
+     
+    </div>
+    <!-- MODAL -->
+
+    <!-- MODAL -->
+    <div class="add-staff" v-if="addServicesProductsDialog">
+        <div class="form-container">
+            <div class="staff-form">
+                <div class="form-input">
+                <div class="d-flex justify-content-between align-items-center">
+                    <h2 v-if="!Update"><i class="fa-solid fa-plus"></i> Add Services Items</h2>
+                    <h2 v-else> <i class="fa-solid fa-edit"></i> Update Services</h2>
+                <!-- <div class="alert alert-danger d-flex flex-column" role="alert"  >
+                    <span v-html="asd"></span>
+                  
+                </div> -->
+                <i class="fa-solid fa-plus add-items-service" @click="addMoreItems()"></i>
+                
+                </div>
+                <form @submit.prevent="addServicesProducts" >
+                <div v-for="(data,i) in ServicesItems" :key="i" class="add-services-products">
+                    <div class="form-group">
+                        <label>Item Name:</label>
+                        <select class="select-dropdown m-0" v-model="ServicesItems[i].product_id">
+                            <option disabled value="">Select Items</option>
+                            <option v-for="data in service.service_product_dropdown" :value="data.id">{{ data.name }}</option>
+                        </select>
+                    </div>
+
+                    <div class="form-group">
+                        <label>Quantity</label>
+                        <input type="text" class="form-control" v-model="ServicesItems[i].quantity" autocomplete="off" @keypress="isNumber($event)" required>
+                        <span  class="text-danger"></span>
+                    </div>
+
+                    <div class="remove-services-product d-flex align-items-center">
+                        <i class="fa-solid fa-minus" @click="removeItem(i)"></i>
+                    </div>
+                 
+                </div>
+               
+               
+                <div v-if="!Update" class="mt-5">
+                    <button type="submit" class="btn btn-primary">Submit</button>
+                    <button type="button" class="btn btn-danger" @click="closeDialog">Cancel</button>
+                </div>
+               
+              </form>
+
+              <div v-if="Update" class="mt-5">
                 <button class="btn btn-primary" @click="submitUpdate()">Update</button>
                 <button type="button" class="btn btn-danger" @click="closeDialog">Cancel</button>
               </div>
@@ -302,5 +448,34 @@ const isNumber = function(evt) {
 
 .table-responsive td{
     padding: 5px;
+}
+.add-items-service{
+    color: black;
+    cursor: pointer;
+}
+.add-services-products{
+    display: flex;
+
+    .form-group{
+        margin: 0px 5px;
+    }
+
+    .form-group:nth-child(1){
+        width: 80%;
+    }
+    .form-group:nth-child(2){
+        width: 20%;
+    }
+
+    .remove-services-product i {
+        background-color: red;
+        padding: 3px;
+        color: white;
+        border-radius: 50%;
+        margin-top: 25px;
+        margin-left: 5px;
+        font-size: 14px;
+        cursor: pointer;
+    }
 }
 </style>
